windows_dev

+++ bin/cmake.sh
@@ -70,6 +70,10 @@ while (($#)); do
             callgrind="--callgrind"
             ;;
 
+        "--no-memory-leak-detection") # [optional] skip runs using valgrind / sanitizer
+            skip="--no-memory-leak-detection"
+            ;;
+
         "--debug") # [optional] runs using debug messaging
             debug="--debug"
             ;;
@@ -104,8 +108,12 @@ fi
 if [[ "${clean}" == "--clean" ]]; then
     if [[ -d "${dir}" ]]; then
         rm -rf "${dir}"
+        mkdir -p "${dir}"
     fi
-    rm -rf "${build}"
+    if [[ -d "${build}" ]]; then
+        rm -rf "${build}"
+        mkdir -p "${build}"
+    fi
 fi
 
 targets=( $(get-source-targets ${source}) )
@@ -125,10 +133,10 @@ if [[ "${cmake}" == "" ]]; then
     exit 8
 fi
 
-[[ ! -d "${build}" ]] && mkdir "${build}"
+[[ ! -d "${build}" ]] && mkdir -p "${build}"
 
 output="${pwd}/output"
-[[ ! -d "${output}" ]] && mkdir "${output}"
+[[ ! -d "${output}" ]] && mkdir -p "${output}"
 
 for target in ${targets[@]}; do
     if [[ -f "${output}/log-${target}.txt" ]]; then
@@ -158,8 +166,8 @@ for config in ${targets[@]}; do
     echo options "$(cmake-options)"
     ${cmake} --build "${build}" --target "${target}" 2>&1 || (echo ERROR: "${target}" && exit 1)
     case "${target}" in
-        c-*) ;& main-*) ;& test-*)
-            if [ ! "$(cat /proc/version | grep -c MSYS)" == "1" ] && [ ! "$(cat /proc/version | grep -c MINGW64)" == "1" ]; then
+        main-*) ;& test-*)
+            if [ ! "${skip}" == "--no-memory-leak-detection" ] && [ ! "$(cat /proc/version | grep -c MSYS)" == "1" ] && [ ! "$(cat /proc/version | grep -c MINGW64)" == "1" ]; then
                 timeout --foreground 180 $(cmake-valgrind-options) "${build}/${target}" 2>&1 >"${output}/log-${target}.txt" || (echo ERROR: "${target}" && exit 1)
             else
                 timeout --foreground 180 "${build}/${target}" 2>&1 >"${output}/log-${target}.txt" || (echo ERROR: "${target}" && exit 1)
